# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ContextOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /runs:
    get:
      operationId: listRuns
      tags: [runs]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Run"
  /runs/{id}:
    get:
      operationId: getRun
      tags: [runs]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunDetail"
  /runs/{id}/pause:
    post:
      operationId: pauseRun
      tags: [runs]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
  /runs/{id}/resume:
    post:
      operationId: resumeRun
      tags: [runs]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
  /runs/{id}/cancel:
    post:
      operationId: cancelRun
      tags: [runs]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
  /runs/{id}/events:
    get:
      operationId: listRunEvents
      tags: [runs]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventLog"
  /actions/{id}:
    get:
      operationId: getAction
      tags: [actions]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Action"
  /approvals:
    get:
      operationId: listApprovals
      tags: [approvals]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApprovalRequest"
  /approvals/{id}/approve:
    post:
      operationId: approveApproval
      tags: [approvals]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRequest"
  /approvals/{id}/deny:
    post:
      operationId: denyApproval
      tags: [approvals]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRequest"
  /artifacts:
    get:
      operationId: listArtifacts
      tags: [artifacts]
      parameters:
        - name: runId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artifact"
  /artifacts/{id}:
    get:
      operationId: getArtifact
      tags: [artifacts]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
  /memory:
    get:
      operationId: listMemory
      tags: [memory]
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: sensitivity
          in: query
          required: false
          schema:
            type: string
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkingMemory"
  /audit:
    get:
      operationId: listAudit
      tags: [audit]
      parameters:
        - name: resourceType
          in: query
          required: false
          schema:
            type: string
        - name: riskTier
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditRecord"
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    JsonObject:
      type: object
      additionalProperties: true
...
    Intent:
      type: object
      required: [id, title, goal, status, riskTier, createdAt]
      properties:
        id: { type: string }
        title: { type: string }
        goal: { type: string }
        constraints: { type: ["string", "null"] }
        successCriteria: { type: ["string", "null"] }
        allowedSystems: { type: ["array", "null"], items: { type: string } }
        deniedSystems: { type: ["array", "null"], items: { type: string } }
        budgetTokens: { type: ["integer", "null"] }
        budgetUsd: { type: ["integer", "null"] }
        maxSteps: { type: ["integer", "null"] }
        riskTier: { type: string }
        status: { type: string }
        runCount: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
...
    Run:
      type: object
      required: [id, intentId, status, orchestrationMode, createdAt]
      properties:
        id: { type: string }
        intentId: { type: string }
        intentTitle: { type: ["string", "null"] }
        status: { type: string }
        orchestrationMode: { type: string }
        leadAgentId: { type: ["string", "null"] }
        summary: { type: ["string", "null"] }
        error: { type: ["string", "null"] }
        tokensUsed: { type: integer }
        costUsdMicros: { type: integer }
        traceId: { type: ["string", "null"] }
        startedAt: { type: ["string", "null"], format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    RunDetail:
      allOf:
        - $ref: "#/components/schemas/Run"
        - type: object
          properties:
            intent: { $ref: "#/components/schemas/Intent" }
            taskGraph: { $ref: "#/components/schemas/JsonObject" }
            actions:
              type: array
              items: { $ref: "#/components/schemas/Action" }
            approvals:
              type: array
              items: { $ref: "#/components/schemas/ApprovalRequest" }
            contextFragments:
              type: array
              items: { $ref: "#/components/schemas/ContextFragment" }
            contextPacks:
              type: array
              items: { $ref: "#/components/schemas/ContextPack" }
            artifacts:
              type: array
              items: { $ref: "#/components/schemas/Artifact" }
            events:
              type: array
              items: { $ref: "#/components/schemas/EventLog" }
            agentRuns:
              type: array
              items: { $ref: "#/components/schemas/AgentRun" }
            agentMessages:
              type: array
              items: { $ref: "#/components/schemas/AgentMessage" }
    Action:
      type: object
      required: [id, runId, name, kind, riskTier, status, createdAt]
      properties:
        id: { type: string }
        runId: { type: string }
        capabilityId: { type: ["string", "null"] }
        nodeId: { type: ["string", "null"] }
        name: { type: string }
        kind: { type: string }
        riskTier: { type: string }
        status: { type: string }
        input: { $ref: "#/components/schemas/JsonObject" }
        output: { $ref: "#/components/schemas/JsonObject" }
        error: { type: ["string", "null"] }
        policyDecision: { $ref: "#/components/schemas/JsonObject" }
        agentId: { type: ["string", "null"] }
        agentRunId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
    ApprovalRequest:
      type: object
      required: [id, runId, actionId, riskTier, status, createdAt]
      properties:
        id: { type: string }
        runId: { type: string }
        actionId: { type: string }
        actionName: { type: ["string", "null"] }
        riskTier: { type: string }
        status: { type: string }
        reason: { type: ["string", "null"] }
        decisionNote: { type: ["string", "null"] }
        decidedAt: { type: ["string", "null"], format: date-time }
        expiresAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    ApprovalDecisionInput:
      type: object
      properties:
        note: { type: string }
    Artifact:
      type: object
      required: [id, name, type, contentType, createdAt]
      properties:
        id: { type: string }
        runId: { type: ["string", "null"] }
        name: { type: string }
        type: { type: string }
        contentType: { type: string }
        content: { type: ["string", "null"] }
        sizeBytes: { type: integer }
        sensitivity: { type: string }
        createdAt: { type: string, format: date-time }
    WorkingMemory:
      type: object
      required: [id, type, key, value, sensitivity, createdAt]
      properties:
        id: { type: string }
        runId: { type: ["string", "null"] }
        type: { type: string }
        key: { type: string }
        value: { type: string }
        sensitivity: { type: string }
        tags: { type: ["array", "null"], items: { type: string } }
        createdAt: { type: string, format: date-time }
    EventLog:
      type: object
      required: [id, type, level, message, createdAt]
      properties:
        id: { type: string }
        runId: { type: ["string", "null"] }
        type: { type: string }
        level: { type: string }
        message: { type: string }
        data: { $ref: "#/components/schemas/JsonObject" }
        agentId: { type: ["string", "null"] }
        agentRunId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    AuditRecord:
      type: object
      required: [id, actorType, action, resourceType, summary, createdAt]
      properties:
        id: { type: string }
        runId: { type: ["string", "null"] }
        actorType: { type: string }
        actorId: { type: ["string", "null"] }
        action: { type: string }
        resourceType: { type: string }
        resourceId: { type: ["string", "null"] }
        summary: { type: string }
        riskTier: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    ContextFragment:
      type: object
      required: [id, type, source, content, tokens, selected, createdAt]
      properties:
        id: { type: string }
        runId: { type: ["string", "null"] }
        type: { type: string }
        source: { type: string }
        content: { type: string }
        tokens: { type: integer }
        relevanceScore: { type: integer }
        selected: { type: boolean }
        rejectionReason: { type: ["string", "null"] }
        sensitivity: { type: string }
        redacted: { type: boolean }
        agentId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    ContextPack:
      type: object
      required: [id, name, totalTokens, strategy, createdAt]
      properties:
        id: { type: string }
        runId: { type: ["string", "null"] }
        name: { type: string }
        fragmentIds: { type: ["array", "null"], items: { type: string } }
        totalTokens: { type: integer }
        strategy: { type: string }
        summary: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
...
    AgentRun:
      type: object
      required: [id, runId, agentId, role, status, createdAt]
      properties:
        id: { type: string }
        runId: { type: string }
        agentId: { type: string }
        agentName: { type: ["string", "null"] }
        parentAgentRunId: { type: ["string", "null"] }
        role: { type: string }
        status: { type: string }
        task: { type: ["string", "null"] }
        output: { $ref: "#/components/schemas/JsonObject" }
        outputValid: { type: ["boolean", "null"] }
        usedFallback: { type: boolean }
        tokensUsed: { type: integer }
        latencyMs: { type: integer }
        costUsdMicros: { type: integer }
        createdAt: { type: string, format: date-time }
    AgentMessage:
      type: object
      required: [id, runId, messageType, content, createdAt]
      properties:
        id: { type: string }
        runId: { type: string }
        fromAgentId: { type: ["string", "null"] }
        toAgentId: { type: ["string", "null"] }
        fromAgentRunId: { type: ["string", "null"] }
        toAgentRunId: { type: ["string", "null"] }
        messageType: { type: string }
        content: { type: string }
        createdAt: { type: string, format: date-time }
...